requirements.txt
transformers
faiss-cpu
pillow
torch
pyautogui
streamlit